requirements.txt
annotated-types==0.6.0
anyio==4.1.0
APScheduler==3.10.4
certifi==2023.11.17
deprecation==2.1.0
exceptiongroup==1.2.0
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
packaging==24.0
postgrest==0.16.3
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2024.1
realtime==1.0.4
six==1.16.0
sniffio==1.3.1
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
typing_extensions==4.11.0
tzlocal==5.2
websockets==12.0
